requirements.txt
Flask>=2.3.0
Werkzeug>=2.3.7
moviepy==1.0.3
Pillow>=11.0.0
numpy
ffmpeg-python>=0.2.0
imageio-ffmpeg==0.4.8
decorator
imageio
setuptools<81
openai>=1.12.0
requests>=2.31.0
elevenlabs>=1.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
psutil>=5.9.0
aiofiles>=23.0.0
pytube>=15.0.0
yt-dlp>=2024.1.0
arabic-reshaper>=3.0.0
python-bidi>=0.4.2
# AI Vocal removal dependencies removed to reduce deployment size
# The following dependencies were removed (saved ~3-4GB):
# demucs>=4.0.0
# torch>=2.0.0  
# torchaudio>=2.0.0
# soundfile>=0.12.1
# replicate>=0.22.0
# audio-separator>=0.17.1
# librosa>=0.10.0
# omegaconf>=2.3.0
# onnxruntime>=1.22.0

# IMPORTANT: After installing dependencies, run the model setup script:
# python3 setup_models.py 
